refactor(api): move share to modules/ as controller, model, service
