+++ apps/api/src/modules/share/index.ts
@@ -1,11 +1,20 @@
 import { Elysia, t } from 'elysia';
-import { noContent } from '../shared/http';
-import { guards, entityGuard } from '../shared/guards';
-import { authContext } from '../shared/auth-context';
-import { HttpError } from '../shared/lib';
-import { commonErrors, errors } from '../shared/responses';
-import { getIssueProjectId } from '../issues/store';
-import { getView } from '../views/store';
+import { noContent } from '#shared/http';
+import { guards, entityGuard } from '#shared/guards';
+import { authContext } from '#shared/auth-context';
+import { HttpError } from '#shared/lib';
+import { commonErrors, errors } from '#shared/responses';
+import { getIssueProjectId } from '../../issues/store';
+import { getView } from '../../views/store';
+import {
+  BundleResponse,
+  ShareTokenResponse,
+  shareExtendedBody,
+  shareIssueParams,
+  shareTokenParams,
+  shareViewParams,
+  sharedViewIssueParams,
+} from './model';
 import {
   enableIssueShare,
   disableIssueShare,
@@ -14,34 +23,7 @@ import {
   getSharedIssue,
   getSharedView,
   getSharedViewIssue,
-} from './store';
-
-// The token is a UUID column; validating its format here turns a malformed token
-// into a 400 rather than letting it reach Postgres as a 500.
-const tokenParams = t.Object({ token: t.String({ format: 'uuid' }) });
-
-// The share link's token, returned when sharing is enabled.
-const ShareTokenResponse = t.Object({ token: t.String() });
-
-// How much the link exposes, sent when enabling it. Enabling an already-shared
-// entity keeps its token and only changes this; leaving the field out keeps that
-// as it stands too, so fetching the link of a live share cannot downgrade it.
-const ShareExtendedBody = t.Optional(
-  t.Object({
-    extended: t.Optional(
-      t.Boolean({
-        description:
-          'Expose the full issues (assignees, labels, custom fields, activity) instead of their title, description, state, type, priority, dates, subtasks and links. Omit to leave it unchanged.',
-      }),
-    ),
-  }),
-);
-
-// The public read-only bundles mirror the store DTOs (project scaffold + entity),
-// which the read-only web pages type themselves. They are self-contained reads,
-// so the response is passed through rather than re-declaring the five feature DTOs
-// they compose.
-const BundleResponse = t.Any();
+} from './service';
 
 export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share'] } })
   .use(authContext)
@@ -69,8 +51,8 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return { token };
     },
     {
-      params: t.Object({ issueId: t.Numeric() }),
-      body: ShareExtendedBody,
+      params: shareIssueParams,
+      body: shareExtendedBody,
       workItem: 'edit',
       response: { 200: ShareTokenResponse, ...commonErrors },
       detail: { summary: 'Enable issue sharing' },
@@ -85,7 +67,7 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return noContent();
     },
     {
-      params: t.Object({ issueId: t.Numeric() }),
+      params: shareIssueParams,
       workItem: 'edit',
       response: { 204: t.Void(), ...commonErrors },
       detail: { summary: 'Revoke issue sharing' },
@@ -100,8 +82,8 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return { token };
     },
     {
-      params: t.Object({ viewId: t.Numeric() }),
-      body: ShareExtendedBody,
+      params: shareViewParams,
+      body: shareExtendedBody,
       savedView: 'edit',
       response: { 200: ShareTokenResponse, ...commonErrors },
       detail: { summary: 'Enable view sharing' },
@@ -116,7 +98,7 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return noContent();
     },
     {
-      params: t.Object({ viewId: t.Numeric() }),
+      params: shareViewParams,
       savedView: 'edit',
       response: { 204: t.Void(), ...commonErrors },
       detail: { summary: 'Revoke view sharing' },
@@ -133,7 +115,7 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return bundle;
     },
     {
-      params: tokenParams,
+      params: shareTokenParams,
       response: { 200: BundleResponse, ...errors(400, 404) },
       detail: { summary: 'Get a shared issue' },
     },
@@ -147,7 +129,7 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return bundle;
     },
     {
-      params: tokenParams,
+      params: shareTokenParams,
       response: { 200: BundleResponse, ...errors(400, 404) },
       detail: { summary: 'Get a shared view' },
     },
@@ -161,7 +143,7 @@ export const shareRoutes = new Elysia({ name: 'share', detail: { tags: ['Share']
       return bundle;
     },
     {
-      params: t.Object({ token: t.String({ format: 'uuid' }), issueId: t.Numeric() }),
+      params: sharedViewIssueParams,
       response: { 200: BundleResponse, ...errors(400, 404) },
       detail: { summary: 'Get an issue from a shared view' },
     },
