+++ apps/api/src/planner.ts
@@ -19,7 +19,7 @@ import { cycleRoutes } from './modules/cycles';
 import { attachmentRoutes } from './modules/attachments';
 import { avatarRoutes } from './modules/avatars';
 import { viewRoutes } from './views/routes';
-import { shareRoutes } from './share/routes';
+import { shareRoutes } from './modules/share';
 import { actionRoutes } from './modules/actions';
 import { webhookRoutes } from './webhooks/routes';
 import { githubSettingsRoutes } from './modules/github';
